Rework twitter scraper, fix media from renamed accounts
Because the graphql endpoint uses user ids it also finds tweets a user
has made with previous names.
Remove a bunch of logic regarding the scrolling back and forth to get more content.
This made everything needlessly complicated and probably didn't even work properly.
Just accept that it is that way and mode on instead.