Simplify submission sample css
Cropper libraries don't play nicely with the previous styling.
At least with this CropperJS is able to properly function, and maybe others too.

Overall the visuals are pretty much the same, except that there's a bit
more padding between the image and file stats. No a bad change though.
It also doesn't waste space when the whole row is smaller than 250px in height.