Fix furrynetwork
Captcha broke. This new endpoint returns NSFW artwork but doesn't paginate.
On top of that, unauthenticated access to characters can break, so in
that case fallback to the same endpoint if neccessary. Luckily that
contains the needed information.

This isn't really that great of a solution, but it works.
On the plus-side, no more selenium.