Simplify the artist/artist url creation code
Arguably one of the worse systems here. I tried to solve some issues with
saving and urls being inserted even though one in the list is invalid.

Of course I didn't add tests so the exact conditions have escaped me since.
Add a few for cases that are definitely relevant (and some others as well).