Handle upcoming `json` gem deprecation
Passing something that is not UTF-8 into it will warn.
As it happens, webmock with httpx returns BINARY unless you
specify the content-type header but I sure as hell won't bother doing that.

Technically this is more correct since in real-life the encoding can't be assumed
to UTF-8 but is probably will be.