Get rid of the horrible pg version check
It just doesn't make a much of sense for what it's doing (and barely works).

Instead, just decouple it from pulling new versions. Upgrading postgres is not such a easy task and it's no big deal
to simply run with an slightly older version.

Of course, this will only work on existing installs if you add the env variable yourself, so show a warning in that case.
When I do the next postgres version udpate, I can simply show a warning that you're using an outdated one. Much simpler.